authentication_service/authentication/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.8
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
python-decouple==3.8
django-cors-headers==4.2.0
factory-boy==3.3.0
django-filter==23.2
gunicorn==21.2.0
toml==0.10.2
black==23.9.1
drf-spectacular==0.26.5
ratelimit==2.2.1
django-debug-toolbar==4.3.0
passlib==1.7.4
Faker==19.6.2
coverage==7.6.10
python-dotenv==1.0.1
requests==2.32.3
